\name{MidasQuantR-package}
\alias{MidasQuantR-package}
\alias{MidasQuantR}
\docType{package}
\title{
\packageTitle{MidasQuantR}
}
\description{
\packageDescription{MidasQuantR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MidasQuantR}
\packageIndices{MidasQuantR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MidasQuantR}

Maintainer: \packageMaintainer{MidasQuantR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
